Move unit tests to test/ directory and add CI workflow for all commits
Test Organization:
- Move unit tests from auth0/tests.py to test/test_backend.py
- Keep functional/e2e tests in test/test_login.py
- Add test/conftest.py for pytest-django configuration
- Clean up auth0/tests.py (now just a comment pointing to test/)

CI/CD Improvements:
- Add new test-ci.yml workflow that runs on ALL branches/commits
  * Runs unit tests on every commit (fast feedback)
  * Only runs functional tests on main branch (requires Auth0 credentials)
- Update python-test-publish.yml to run both unit and functional tests
  * Unit tests run with mock credentials
  * Functional tests run with real Auth0 credentials

Dependencies:
- Add pytest-django>=4.0.0 to test dependencies
- Configure pytest to work with Django tests properly
- Remove --browser chromium from default pytest options (only for e2e)

All 17 unit tests passing successfully.
