+++ .github/workflows/test-ci.yml
@@ -1,0 +1,71 @@
+name: Test CI
+
+on:
+  push:
+    branches: ["**"]  # Run on all branches
+  pull_request:
+    branches: ["**"]  # Run on PRs to any branch
+
+jobs:
+  unit-tests:
+    name: Unit Tests
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.11", "3.12", "3.13"]
+
+    steps:
+    - uses: actions/checkout@v4
+
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.python-version }}
+
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -e ".[test]"
+
+    - name: Run unit tests
+      env:
+        AUTH0_CLIENT_ID: test-client-id
+        AUTH0_CLIENT_SECRET: test-client-secret
+        AUTH0_DOMAIN: test.auth0.com
+        AUTH0_AUDIENCE: test-audience
+      run: |
+        pytest test/test_backend.py -v
+
+  functional-tests:
+    name: Functional Tests (E2E)
+    runs-on: ubuntu-latest
+    # Only run functional tests if Auth0 credentials are available
+    if: github.event_name == 'push' && (github.ref == 'refs/heads/main' || startsWith(github.ref, 'refs/tags/'))
+    strategy:
+      matrix:
+        python-version: ["3.11", "3.12", "3.13"]
+
+    steps:
+    - uses: actions/checkout@v4
+
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.python-version }}
+
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -e ".[test]"
+        playwright install chromium
+
+    - name: Run functional tests
+      env:
+        AUTH0_USERNAME: ${{ secrets.AUTH0_USERNAME }}
+        AUTH0_PASSWORD: ${{ secrets.AUTH0_PASSWORD }}
+        AUTH0_CLIENT_ID: ${{ secrets.AUTH0_CLIENT_ID }}
+        AUTH0_CLIENT_SECRET: ${{ secrets.AUTH0_CLIENT_SECRET }}
+        AUTH0_DOMAIN: ${{ secrets.AUTH0_DOMAIN }}
+        AUTH0_AUDIENCE: ${{ secrets.AUTH0_AUDIENCE }}
+      run: |
+        pytest test/test_login.py -v -m e2e --browser chromium
